Reshape Summary to match Java's ResultSummary surface
The previous `Summary` class exposed a public `attr_reader :metadata`
and bundled inner classes for Query/ServerInfo/etc. — neither matches
the Java driver's `org.neo4j.driver.summary.ResultSummary` API and
forced testkit-backend to reach into raw wire metadata.

Convert `Summary` into a namespace module containing one class per
Java type (`ResultSummary`, `Query`, `ServerInfo`, `DatabaseInfo`,
`SummaryCounters`, `Plan`, `Profile`, `Notification`, `QueryType`),
each in its own file under `lib/mri/neo4j/driver/summary/`. Zeitwerk
autovivifies the module from the directory.

`ResultSummary#metadata` is now private. The single internal consumer
(`Session#harvest_auto_commit_bookmark`) goes through `send(:metadata)`
with a comment explaining why — Java's ResultSummary doesn't expose a
bookmark accessor either, since it only ever lives on the wire.

Add `#to_h` to `Notification`, `Plan`, and `Profile` (camelCase string
keys, matching Bolt wire shape). On the JRuby side, prepend matching
`#to_h` onto `InternalNotification`, `InternalPlan`, and the new
`InternalProfiledPlan` extension so testkit-backend's payload code is
flavour-agnostic.

Rewrite `testkit-backend/summary_payload.rb` to use only the public
Summary API: typed `query_type` mapped back to the wire token via a
small lookup, `notifications.map(&:to_h)`, `plan&.to_h`, `profile&.to_h`,
`counters.public_send(...)`. No more `summary.metadata[...]` reaches.

Result: testkit `tests.stub.basic_query.test_basic_query` passes
cleanly on MRI (3 pass / 4 skip — matches baseline) and produces the
expected output on JRuby (1 pre-existing JRuby failure unrelated to
this refactor). Full `bundle exec rspec` is green (400 examples,
0 failures) on MRI.

Investigated the JRuby stub-timeout issue separately: confirmed
retry-based tests (e.g. `test_should_retry_write_tx_until_success_on_run_error`)
take ~30s each because the driver's `maxTransactionRetryTime` default
is exhausted against stubs that never recover. The fix likely lives in
per-test timeout-config plumbing rather than the driver itself, and is
out of scope for this PR — left as follow-up.
