+++ lib/jruby/neo4j/driver/ext/internal/summary/internal_notification.rb
@@ -19,6 +19,25 @@ module Neo4j
             def category
               super.or_else(nil)
             end
+
+            # camelCase string-keyed serialisation, matching MRI's
+            # Neo4j::Driver::Summary::Notification#to_h. testkit-backend's
+            # SummaryPayload calls this so the dict shape is identical
+            # whether `summary` is an MRI ResultSummary or a Java one.
+            def to_h
+              {
+                'code' => code,
+                'title' => title,
+                'description' => description,
+                'severityLevel' => severity_level,
+                'category' => category,
+                'position' => position && {
+                  'offset' => position.offset,
+                  'line' => position.line,
+                  'column' => position.column
+                }
+              }.compact
+            end
           end
         end
       end
