+++ lib/jruby/neo4j/driver/ext/internal/summary/internal_plan.rb
@@ -14,9 +14,17 @@ module Neo4j
               super.to_a
             end
 
-            # this part should probably be included in testkit-backend
+            # camelCase string-keyed serialisation, matching MRI's
+            # Neo4j::Driver::Summary::Plan#to_h. testkit-backend's
+            # SummaryPayload uses it directly so the dict shape is
+            # identical whether `summary.plan` is an MRI Plan or a Java one.
             def to_h
-              {operator_type:, args:, identifiers:, children: children&.map(&:to_h)}
+              {
+                'operatorType' => operator_type,
+                'identifiers' => identifiers,
+                'args' => args,
+                'children' => children&.map(&:to_h)
+              }
             end
           end
         end
